requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dateparser==1.2.0
dnspython==2.7.0
email_validator==2.2.0
eventlet==0.37.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
psycopg2-binary==2.9.10
pycryptodome==3.21.0
python-binance==1.0.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.36
stripe==11.2.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.2
wsproto==1.2.0
WTForms==3.2.1
yarl==1.17.1
